Modifications to dashboard charts
Also applied temporary fix to the reauthenticatingFetch() function - this function still doesn't refresh the tokens correctly.
